Partie_2/OpenAi/Exercice/requirement.txt
openai==1.54.0
python-dotenv==1.0.0